Fix SignalR NotificationHub UserIdentifier error and add deployment config
- Added ClaimTypes.NameIdentifier and Sub claims to JWT tokens for SignalR compatibility
- Fixed NotificationHub to extract user ID from claims instead of Context.UserIdentifier
- Added graceful error handling for missing user claims
- Added Railway deployment configuration (railway.json, nixpacks.toml)
- Added Docker deployment files (Dockerfile, docker-compose.yml)
- Added production appsettings template and database deployment script

+++ src/HallApp.Application/Services/HallService.cs
@@ -87,7 +87,7 @@ public class HallService : IHallService
         return new List<Hall>();
     }
 
-    public async Task<List<Hall>> GetHallsByPriceRangeAsync(decimal minPrice, decimal maxPrice)
+    public async Task<List<Hall>> GetHallsByPriceRangeAsync(double minPrice, double maxPrice)
     {
         // Placeholder - implementation depends on actual repository methods
         return new List<Hall>();
